+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.x.x Unreleased
 ### New features
+* [experimental] Enable dask chunking information to be passed to `InferenceData.from_netcdf` with regex support ([1749](https://github.com/arviz-devs/arviz/pull/1749))
 
 ### Maintenance and fixes
 * Bokeh 3 compatibility. ([1919](https://github.com/arviz-devs/arviz/pull/1919))
@@ -32,9 +33,6 @@
 * Added ability to plot HDI contours to `plot_kde` with the new `hdi_probs` parameter. ([1665](https://github.com/arviz-devs/arviz/pull/1665))
 * Add dtype parsing and setting in all Stan converters ([1632](https://github.com/arviz-devs/arviz/pull/1632))
 * Add option to specify colors for each element in ppc_plot  ([1769](https://github.com/arviz-devs/arviz/pull/1769))
-* Enable dask chunking information to be passed to `InferenceData.from_netcdf` ([1749](https://github.com/arviz-devs/arviz/pull/1749))
-* Enable dask chunking information to be passed to `InferenceData.from_netcdf` with regex support ([1749](https://github.com/arviz-devs/arviz/pull/1749))
-* Enable dask chunking information to be passed to `InferenceData.from_netcdf` with regex support to enable parallel io ([1749](https://github.com/arviz-devs/arviz/pull/1749))
 
 ### Maintenance and fixes
 * Fix conversion for numpyro models with ImproperUniform latent sites ([1713](https://github.com/arviz-devs/arviz/pull/1713))